check for ~/minecraft_server, make sure no sudo is used, use mojang download link

+++ setup.sh
@@ -2,13 +2,19 @@ sudo apt install update
 
 sudo apt install wget openjdk-8-jdk screen -y
 
+if [ -e ~/minecraft_server ]; then
+    echo
+    echo "$HOME/minecraft_server exists! Please delete it first!"
+    exit
+fi
+
 mkdir ~/minecraft_server
 rm ~/minecraft_server/setup.sh
 rm ~/minecraft_server/run.sh
 cp setup.sh ~/minecraft_server
 cp run.sh ~/minecraft_server
 cd ~/minecraft_server
-wget -O minecraft_server.jar https://s3.amazonaws.com/Minecraft.Download/versions/1.11.2/minecraft_server.1.11.2.jar
+wget -O minecraft_server.jar https://launcher.mojang.com/v1/objects/a412fd69db1f81db3f511c1463fd304675244077/server.jar
 chmod +x minecraft_server.jar
 echo eula=true >> eula.txt
 ./run.sh
